Rename Wsgi* to Uwsgi*, wsgi.h/.cpp to uwsgi.h/.cpp
